clean up AST and type printers
