+++ pynestml/codegeneration/latex_expression_printer.py
@@ -31,7 +31,7 @@ from pynestml.utils.ast_utils import ASTUtils
 
 
 class LatexExpressionPrinter(Printer):
-    """
+    r"""
     Pretty printer for LaTeX. Assumes to be printing in a LaTeX environment where math mode is already on.
     """
 
@@ -114,17 +114,18 @@ class LatexExpressionPrinter(Printer):
 
             raise Exception("Unknown node type")
 
-        raise RuntimeError("Unsupported rhs in rhs pretty printer!")
+        raise RuntimeError("Tried to print unknown expression: \"%s\"" % str(node))
 
     def print_function_call(self, function_call: ASTFunctionCall) -> str:
         function_name = self.reference_converter.convert_function_call(function_call)
         if ASTUtils.needs_arguments(function_call):
             return function_name % self.print_function_call_argument_list(function_call)
-        else:
-            return function_name
+
+        return function_name
 
     def print_function_call_argument_list(self, function_call: ASTFunctionCall) -> Tuple[str, ...]:
         ret = []
         for arg in function_call.get_args():
             ret.append(self.print_expression(arg))
+
         return tuple(ret)
