+++ pynestml/codegeneration/nest_reference_converter.py
@@ -19,21 +19,14 @@
 # You should have received a copy of the GNU General Public License
 # along with NEST.  If not, see <http://www.gnu.org/licenses/>.
 
-from typing import Union
-
 import re
 
-from pynestml.codegeneration.base_reference_converter import BaseReferenceConverter
-from pynestml.codegeneration.reference_converter import ReferenceConverter
+from pynestml.codegeneration.cpp_reference_converter import CppReferenceConverter
 from pynestml.codegeneration.unit_converter import UnitConverter
-from pynestml.meta_model.ast_arithmetic_operator import ASTArithmeticOperator
-from pynestml.meta_model.ast_bit_operator import ASTBitOperator
-from pynestml.meta_model.ast_comparison_operator import ASTComparisonOperator
-from pynestml.meta_model.ast_function_call import ASTFunctionCall
-from pynestml.meta_model.ast_logical_operator import ASTLogicalOperator
-from pynestml.meta_model.ast_unary_operator import ASTUnaryOperator
 from pynestml.meta_model.ast_variable import ASTVariable
 from pynestml.meta_model.ast_external_variable import ASTExternalVariable
+from pynestml.meta_model.ast_function_call import ASTFunctionCall
+from pynestml.symbol_table.scope import Scope, ScopeType
 from pynestml.symbols.predefined_functions import PredefinedFunctions
 from pynestml.symbols.predefined_units import PredefinedUnits
 from pynestml.symbols.predefined_variables import PredefinedVariables
@@ -46,48 +39,27 @@ from pynestml.utils.logger import Logger, LoggingLevel
 from pynestml.utils.messages import Messages
 
 
-class NESTReferenceConverter(BaseReferenceConverter):
+class NESTReferenceConverter(CppReferenceConverter):
     r"""
-    This concrete reference converter is used to transfer internal names to counter-pieces in NEST.
+    Reference converter for C++ syntax and using the NEST API.
     """
 
-    def convert_binary_op(self, binary_operator) -> str:
-        """
-        Converts a single binary operator to nest processable format.
-        :param binary_operator: a single binary operator string.
-        :type binary_operator: AST_
-        :return: the corresponding nest representation
-        """
-        if isinstance(binary_operator, ASTArithmeticOperator):
-            return self.convert_arithmetic_operator(binary_operator)
-
-        if isinstance(binary_operator, ASTBitOperator):
-            return self.convert_bit_operator(binary_operator)
-
-        if isinstance(binary_operator, ASTComparisonOperator):
-            return self.convert_comparison_operator(binary_operator)
-
-        if isinstance(binary_operator, ASTLogicalOperator):
-            return self.convert_logical_operator(binary_operator)
-
-        raise RuntimeError('Cannot determine binary operator!')
-
-    def convert_function_call(self, function_call, prefix='') -> str:
+    def convert_function_call(self, function_call: ASTFunctionCall, prefix: str = '') -> str:
         """
         Converts a single handed over function call to C++ NEST API syntax.
 
         Parameters
         ----------
-        function_call : ASTFunctionCall
+        function_call
             The function call node to convert.
-        prefix : str
+        prefix
             Optional string that will be prefixed to the function call. For example, to refer to a function call in the class "node", use a prefix equal to "node." or "node->".
 
             Predefined functions will not be prefixed.
 
         Returns
         -------
-        s : str
+        s
             The function call string in C++ syntax.
         """
         function_name = function_call.get_name()
@@ -178,11 +150,10 @@ e();
             return prefix + function_name + '(' + ', '.join(['{!s}' for _ in range(n_args)]) + ')'
         return prefix + function_name + '()'
 
-    def convert_name_reference(self, variable: ASTVariable, prefix='') -> str:
+    def convert_name_reference(self, variable: ASTVariable, prefix: str = '') -> str:
         """
         Converts a single variable to nest processable format.
         :param variable: a single variable.
-        :type variable: ASTVariable
         :return: a nest processable format.
         """
         if isinstance(variable, ASTExternalVariable):
@@ -209,6 +180,7 @@ e();
 
         if symbol.is_local():
             return variable.get_name() + ('[' + variable.get_vector_parameter() + ']' if symbol.has_vector_parameter() else '')
+
         if symbol.is_buffer():
             if isinstance(symbol.get_type_symbol(), UnitTypeSymbol):
                 units_conversion_factor = UnitConverter.get_factor(symbol.get_type_symbol().unit.unit)
@@ -227,8 +199,7 @@ e();
         if symbol.is_inline_expression:
             return 'get_' + variable.get_name() + '()' + ('[i]' if symbol.has_vector_parameter() else '')
 
-        if symbol.is_kernel():
-            assert False, "NEST reference converter cannot print kernel; kernel should have been converted during code generation"
+        assert not symbol.is_kernel(), "NEST reference converter cannot print kernel; kernel should have been converted during code generation"
 
         if symbol.is_state():
             temp = ""
@@ -247,10 +218,7 @@ e();
             self.name(symbol) + \
             ('[' + variable.get_vector_parameter() + ']' if symbol.has_vector_parameter() else '')
 
-    def __get_unit_name(self, variable):
-        assert (variable is not None and isinstance(variable, ASTVariable)), \
-            '(PyNestML.CodeGeneration.NestReferenceConverter) No or wrong type of uses-gsl provided (%s)!' % type(
-                variable)
+    def __get_unit_name(self, variable: ASTVariable):
         assert variable.get_scope() is not None, "Undeclared variable: " + variable.get_complete_name()
 
         variable_name = self.convert_to_cpp_name(variable.get_complete_name())
@@ -260,11 +228,10 @@ e();
 
         return ''
 
-    def convert_print_statement(self, function_call) -> str:
-        """
+    def convert_print_statement(self, function_call: ASTFunctionCall) -> str:
+        r"""
         A wrapper function to convert arguments of a print or println functions
         :param function_call: print function call
-        :type function_call: ASTFunctionCall
         :return: the converted print string with corresponding variables, if any
         """
         stmt = function_call.get_args()[0].get_string()
@@ -272,8 +239,8 @@ e();
         scope = function_call.get_scope()
         return self.__convert_print_statement_str(stmt, scope)
 
-    def __convert_print_statement_str(self, stmt, scope) -> str:
-        """
+    def __convert_print_statement_str(self, stmt: str, scope: Scope) -> str:
+        r"""
         Converts the string argument of the print or println function to NEST processable format
         Variables are resolved to NEST processable format and printed with physical units as mentioned in model, separated by a space
 
@@ -294,9 +261,7 @@ e();
             std::cout << "Membrane potential = " << V_m << " mV";
 
         :param stmt: argument to the print or println function
-        :type stmt: str
         :param scope: scope of the variables in the argument, if any
-        :type scope: Scope
         :return: the converted string to NEST
         """
         pattern = re.compile(r'\{[a-zA-Z_][a-zA-Z0-9_]*\}')  # Match the variables enclosed within '{ }'
@@ -309,160 +274,8 @@ e();
             ast_var = ASTVariable(var_name, scope=scope)
             right = ' ' + self.__get_unit_name(ast_var) + right  # concatenate unit separated by a space with the right part of the string
             return fun_left(left) + self.convert_name_reference(ast_var) + fun_right(right)
-        else:
-            return '"' + stmt + '"'  # format bare string in C++ (add double quotes)
 
-    def convert_constant(self, const: Union[str, float, int]) -> str:
-        """
-        Converts a single handed over constant.
-        :param constant_name: a constant as string.
-        :type constant_name: str
-        :return: the corresponding nest representation
-        """
-        if const == 'inf':
-            return 'std::numeric_limits<double_t>::infinity()'
-
-        if const == 'true':
-            return 'true'
-
-        if const == 'false':
-            return 'false'
-
-        if isinstance(const, float) or isinstance(const, int):
-            return str(const)
-
-        return const
-
-    def convert_unary_op(self, unary_operator) -> str:
-        """
-        Depending on the concretely used operator, a string is returned.
-        :param unary_operator: a single operator.
-        :type unary_operator:  ASTUnaryOperator
-        :return: the same operator
-        """
-        if unary_operator.is_unary_plus:
-            return '(' + '+' + '%s' + ')'
-
-        if unary_operator.is_unary_minus:
-            return '(' + '-' + '%s' + ')'
-
-        if unary_operator.is_unary_tilde:
-            return '(' + '~' + '%s' + ')'
-
-        raise RuntimeError('Cannot determine unary operator!', LoggingLevel.ERROR)
-
-    def convert_encapsulated(self) -> str:
-        """
-        Converts the encapsulating parenthesis to NEST style.
-        :return: a set of parenthesis
-        """
-        return '(%s)'
-
-    def convert_logical_not(self) -> str:
-        """
-        Returns a representation of the logical not in NEST.
-        :return: a string representation
-        """
-        return '(' + '!' + '%s' + ')'
-
-    def convert_logical_operator(self, op) -> str:
-        """
-        Prints a logical operator in NEST syntax.
-        :param op: a logical operator object
-        :type op: ASTLogicalOperator
-        :return: a string representation
-        """
-        if op.is_logical_and:
-            return '%s' + '&&' + '%s'
-        elif op.is_logical_or:
-            return '%s' + '||' + '%s'
-        else:
-            raise RuntimeError('Cannot determine logical operator!', LoggingLevel.ERROR)
-
-    def convert_comparison_operator(self, op) -> str:
-        """
-        Prints a logical operator in NEST syntax.
-        :param op: a logical operator object
-        :type op: ASTComparisonOperator
-        :return: a string representation
-        """
-        if op.is_lt:
-            return '%s' + '<' + '%s'
-
-        if op.is_le:
-            return '%s' + '<=' + '%s'
-
-        if op.is_eq:
-            return '%s' + '==' + '%s'
-
-        if op.is_ne or op.is_ne2:
-            return '%s' + '!=' + '%s'
-
-        if op.is_ge:
-            return '%s' + '>=' + '%s'
-
-        if op.is_gt:
-            return '%s' + '>' + '%s'
-
-        raise RuntimeError('Cannot determine comparison operator!')
-
-    def convert_bit_operator(self, op) -> str:
-        """
-        Prints a logical operator in NEST syntax.
-        :param op: a logical operator object
-        :type op: ASTBitOperator
-        :return: a string representation
-        """
-        if op.is_bit_shift_left:
-            return '%s' + '<<' '%s'
-
-        if op.is_bit_shift_right:
-            return '%s' + '>>' + '%s'
-
-        if op.is_bit_and:
-            return '%s' + '&' + '%s'
-
-        if op.is_bit_or:
-            return '%s' + '|' + '%s'
-
-        if op.is_bit_xor:
-            return '%s' + '^' + '%s'
-
-        raise RuntimeError('Cannot determine bit operator!')
-
-    def convert_arithmetic_operator(self, op) -> str:
-        """
-        Prints a logical operator in NEST syntax.
-        :param op: a logical operator object
-        :type op: ASTArithmeticOperator
-        :return: a string representation
-        """
-        if op.is_plus_op:
-            return '%s' + ' + ' + '%s'
-
-        if op.is_minus_op:
-            return '%s' + ' - ' + '%s'
-
-        if op.is_times_op:
-            return '%s' + ' * ' + '%s'
-
-        if op.is_div_op:
-            return '%s' + ' / ' + '%s'
-
-        if op.is_modulo_op:
-            return '%s' + ' %% ' + '%s'
-
-        if op.is_pow_op:
-            return 'pow' + '(%s, %s)'
-
-        raise RuntimeError('Cannot determine arithmetic operator!')
-
-    def convert_ternary_operator(self) -> str:
-        """
-        Prints a ternary operator in NEST syntax.
-        :return: a string representation
-        """
-        return '(' + '%s' + ') ? (' + '%s' + ') : (' + '%s' + ')'
+        return '"' + stmt + '"'  # format bare string in C++ (add double quotes)
 
     def print_origin(self, variable_symbol: VariableSymbol, prefix: str = '') -> str:
         """
