+++ pynestml/codegeneration/unitless_expression_printer.py
@@ -20,9 +20,7 @@
 # along with NEST.  If not, see <http://www.gnu.org/licenses/>.
 
 from pynestml.codegeneration.expressions_printer import ExpressionsPrinter
-from pynestml.codegeneration.reference_converter import ReferenceConverter
-from pynestml.codegeneration.nestml_reference_converter import NestMLReferenceConverter
-from pynestml.meta_model.ast_expression import ASTExpression
+from pynestml.meta_model.ast_expression_node import ASTExpressionNode
 from pynestml.meta_model.ast_simple_expression import ASTSimpleExpression
 from pynestml.symbols.symbol import SymbolKind
 from pynestml.symbols.predefined_units import PredefinedUnits
@@ -30,30 +28,30 @@ from pynestml.codegeneration.unit_converter import UnitConverter
 
 
 class UnitlessExpressionPrinter(ExpressionsPrinter):
-    """
+    r"""
     An adjusted version of the printer which does not print units with literals.
     """
 
-    def print_expression(self, node, prefix=''):
-        """Print an expression.
+    def print_expression(self, node: ASTExpressionNode, prefix: str = '') -> str:
+        r"""Print an expression.
 
         Parameters
         ----------
-        node : ASTExpressionNode
+        node
             The expression node to print.
-        prefix : str
+        prefix
             *See documentation for the function ExpressionsPrinter::print_function_call().*
-
 
         Returns
         -------
-        s : str
+        s
             The expression string.
         """
         if isinstance(node, ASTSimpleExpression):
             if node.is_numeric_literal():
                 return self.reference_converter.convert_constant(node.get_numeric_literal())
-            elif node.is_variable() and node.get_scope() is not None:
+
+            if node.is_variable() and node.get_scope() is not None:
                 node_is_variable_symbol = node.get_scope().resolve_to_symbol(
                     node.variable.get_complete_name(), SymbolKind.VARIABLE) is not None
                 if not node_is_variable_symbol and PredefinedUnits.is_unit(node.variable.get_complete_name()):
